Refactor: 신고 도메인 패키지 구조 리팩토링
